Move GenericTable
